Renames About component to index.js
